requirements.txt
openai>=1.12.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
chromadb>=0.4.0
jinja2>=3.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
gitpython>=3.1.0
watchdog>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
